# swagger/swagger.yaml
openapi: 3.0.0
info:
  title: HERMES API Docs
  version: 1.0.0
  description: 라즈베리파이 기반 스마트 알림 시스템 API 문서
servers:
  - url: http://localhost:8080

paths:
  /app-notif:
    post:
      summary: 알림 등록
      description: 앱에서 마스킹된 알림을 서버로 전송 후 요약 저장
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                device_id:
                  type: string
                  example: phone001
                device_type:
                  type: string
                  example: android
                Message:
                  type: object
                  properties:
                    app_package:
                      type: string
                      example: com.kakao.talk
                    title:
                      type: string
                      example: 홍길동
                    text:
                      type: string
                      example: 내일 2시에 봐요
                    posted_at_utc:
                      type: string
                      example: 2025-08-19T03:45:12Z
      responses:
        200:
          description: 알림 저장 성공

  /app-notif:
    get:
      summary: 알림 목록 조회
      description: DB에 저장된 알림들을 반환합니다.
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            example: 10
      responses:
        200:
          description: 조회 성공

  /app-notif/{id}:
    delete:
      summary: 특정 알림 삭제
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            example: 3
      responses:
        200:
          description: 삭제 성공

  /app-notif/speak-all:
    post:
      summary: 전체 알림을 스피커로 전송
      description: DB의 모든 요약 알림을 합쳐 MQTT로 전송
      responses:
        200:
          description: 전송 완료

  /speaker/volume:
    post:
      summary: 스피커 볼륨 조절
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                volume:
                  type: integer
                  example: 85
      responses:
        200:
          description: MQTT로 볼륨 전송 성공